fix(types): use f64 for GetMiningInfo networkhashps field
Bitcoin Core returns networkhashps as a double. The previous type (i64)
causes deserialisation to fail.

+++ types/src/model/mining.rs
@@ -113,7 +113,7 @@ pub struct GetMiningInfo {
     /// The current target (v29 onwards).
     pub target: Option<Target>,
     /// The network hashes per second.
-    pub network_hash_ps: i64,
+    pub network_hash_ps: f64,
     /// The size of the mempool.
     pub pooled_tx: i64,
     /// Minimum feerate of packages selected for block inclusion.
